Fix negative-coordinate truncation, hanging-entity gaps, and API semantics
Third audit pass over the previously unread files (mixins, map
integrations, public API, flag types):

- HangingEntityMixin: add the missing per-world useRegions kill-switch
  check (the only mixin without it), and replace the plain (int) casts
  with Math.floor — truncation toward zero mis-binned frames/paintings
  by one block at negative coordinates, so region-border decorations in
  negative quadrants escaped protection.
- BlockEventHandler.onExplosion: same truncation bug for decorations in
  the affected-entities filter; pass raw doubles to testState instead.
- WorldGuardNeoAPI: canBuild/canPlace/canInteract/canAccessChests are
  documented as equivalent to the event-handler checks but used plain
  testState, which answers ALLOW for foreign claims with no explicit
  flags (no implicit membership protection). Switched to testBuildAccess
  so third-party mods get the same verdict the handlers enforce.

Verified: gradle build passes.

+++ src/main/java/dev/thefather007/worldguardneo/listeners/BlockEventHandler.java
@@ -258,8 +258,9 @@ public final class BlockEventHandler {
                 // the global / region-level flag value matters.
                 if (ent instanceof net.minecraft.world.entity.decoration.HangingEntity
                         || ent instanceof net.minecraft.world.entity.decoration.ArmorStand) {
-                    return !mgr.testState(finalFlag, null,
-                            (int) ent.getX(), (int) ent.getY(), (int) ent.getZ());
+                    // Pass raw doubles — testState takes doubles; the old (int) casts truncated
+                    // toward zero, mis-binning decorations at negative coordinates by one block.
+                    return !mgr.testState(finalFlag, null, ent.getX(), ent.getY(), ent.getZ());
                 }
                 return false; // Mobs, items, projectiles etc. — vanilla rules.
             });
